requirements.txt
Flask
Flask-SQLAlchemy
requests
python-dotenv
gunicorn
uuid
psycopg2
Flask-Migrate
